+++ packages/react/src/Pagehead/Pagehead.dev.stories.tsx
@@ -1,4 +1,4 @@
-import type {Meta} from '@storybook/react'
+import type {Meta} from '@storybook/react-vite'
 import Pagehead from './Pagehead'
 import Heading from '../Heading'