+++ packages/react/src/SegmentedControl/SegmentedControlButton.stories.tsx
@@ -1,4 +1,4 @@
-import type {Meta, StoryFn} from '@storybook/react'
+import type {Meta, StoryFn} from '@storybook/react-vite'
 import type {SegmentedControlButtonProps} from './SegmentedControlButton'
 import SegmentedControlButton from './SegmentedControlButton'
 import {SegmentedControl} from '.'