+++ packages/react/src/TreeView/TreeView.features.stories.tsx
@@ -6,7 +6,7 @@ import {
   FileIcon,
   KebabHorizontalIcon,
 } from '@primer/octicons-react'
-import type {Meta, StoryFn} from '@storybook/react'
+import type {Meta, StoryFn} from '@storybook/react-vite'
 import React from 'react'
 import Box from '../Box'
 import {Button, IconButton} from '../Button'