+++ packages/react/src/UnderlineNav/UnderlineNav.tsx
@@ -32,6 +32,12 @@ export type UnderlineNavProps = {
    * loading state for all counters. It displays loading animation for individual counters (UnderlineNav.Item) until all are resolved. It is needed to prevent multiple layout shift.
    */
   loadingCounters?: boolean
+  /**
+   * There are cases where you may not want the horizontal padding on items,
+   * and panels to make the tabs look horizontally aligned with the content above and below it.
+   * Setting this to `flush` will remove the horizontal padding on the items.
+   */
+  variant?: 'inset' | 'flush'
 }
 // When page is loaded, we don't have ref for the more button as it is not on the DOM yet.
 // However, we need to calculate number of possible items when the more button present as well. So using the width of the more button as a constant.
@@ -141,6 +147,7 @@ export const UnderlineNav = forwardRef(
       'aria-label': ariaLabel,
       sx: sxProp = defaultSxProp,
       loadingCounters = false,
+      variant = 'inset',
       className,
       children,
     }: UnderlineNavProps,
@@ -182,6 +189,7 @@ export const UnderlineNav = forwardRef(
 
     if (__DEV__) {
       // Practically, this is not a conditional hook, it is just making sure this hook runs only on DEV not PROD.
+      // eslint-disable-next-line react-compiler/react-compiler
       // eslint-disable-next-line react-hooks/rules-of-hooks
       useEffect(() => {
         // Address illegal state where there are multiple items that have `aria-current='page'` attribute
@@ -308,7 +316,14 @@ export const UnderlineNav = forwardRef(
         }}
       >
         {ariaLabel && <VisuallyHidden as="h2">{`${ariaLabel} navigation`}</VisuallyHidden>}
-        <UnderlineWrapper as={as} aria-label={ariaLabel} className={className} ref={navRef} sx={sxProp}>
+        <UnderlineWrapper
+          as={as}
+          aria-label={ariaLabel}
+          className={className}
+          ref={navRef}
+          sx={sxProp}
+          data-variant={variant}
+        >
           <UnderlineItemList ref={listRef} role="list">
             {listItems}
             {menuItems.length > 0 && (